requirements.txt
requests
pandas
numpy
scikit-learn
xgboost
joblib
plotly
python-dotenv
google-genai
pydantic